Split types into own file
