+++ src/browser.ts
@@ -1,24 +1,12 @@
 // Browser-only exports. Must not import Node-only modules
 
 export {
-  Aggregate,
-  AggregateMatch,
-  AggregateMatches,
   Aggregator,
-  CacheAdapter,
-  DuplicateError,
-  EmitterAdapter,
   Event,
   EventContext,
   EventData,
-  EventHandler,
   EventNamespace,
-  EventStore,
   EventType,
-  ExecuteF,
   Logger,
-  StoreAdapter,
-  Subscriptions,
-  ValidateF,
-} from ".";
+} from "./types";
 export { createEvent, createContext, isEventData, isEventContext, isEventType } from "./helpers";
